Minor correction after core review.
Additional test simplification.
Now it needs only two connections,
where 1st fails asap
and 2nd fails with delay - giving connection reaper time to meet deadlock
